Update test for PR #3457
