+++ tests/binaries/host/stack_vars.native.c
@@ -1,0 +1,28 @@
+#include <stdio.h>
+#include <string.h>
+
+// Portable breakpoint
+void break_here(void) {}
+
+void inner_function(void) {
+    char buffer[64];
+    int local_var = 42;
+
+    strcpy(buffer, "Hello from inner function!");
+
+    printf("local_var = %d\n", local_var);
+    printf("buffer = %s\n", buffer);
+
+    break_here();
+}
+
+void outer_function() {
+    inner_function();
+}
+
+int main() {
+    printf("Starting test program...\n");
+    outer_function();
+    printf("Done!\n");
+    return 0;
+}
